fix: 도서 점검(Checklist) 유효성 검증 오류 해결 및 DBML 스키마 동기화

+++ src/main/java/com/example/yoohoo_be/checklists/dto/BookCheckCompletedListResponseDto.java
@@ -11,18 +11,18 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class BookCheckCompletedListResponseDto {
 
-    private Long resultBatchId;
-    private Long bookId;
-    private String title;
-    private String author;
-    private String publisher;
-    private String isbn;
-    private String kdcCode;    // [수정됨] genre -> kdcCode (실제 DB 필드로 교체)
-    private String kdcClass;   // [수정됨] turnoverRate -> kdcClass (실제 DB 필드로 교체)
-    private String callNumber;
-    private String coverUrl;
-    private String checkedDate;
-    private String librarianName;
-    private Integer totalScore;
-    private String status;
+    private Long resultBatchId;   // 점검 제출 그룹 ID (BookCheckBatch PK)
+    private Integer bookId;          // 도서 고유 ID
+    private String title;         // 도서 제목
+    private String author;        // 저자
+    private String publisher;       // 출판사
+    private String isbn;            // 도서 ISBN
+    private String kdcCode;         // KDC 코드
+    private String kdcClass;        // KDC 분류기호
+    private String callNumber;      // 청구기호
+    private String coverUrl;        // 도서 표지 이미지 URL
+    private String checkedDate;     // 점검 날짜 (YYYY-MM-DD)
+    private String librarianName; // 담당 사서 코드(또는 이름)
+    private Integer totalScore;   // 총점
+    private String status;        // 도서 상태
 }
